backend/server/requirements.txt
alembic==1.14.1
aniso8601==10.0.0
asttokens==3.0.0
backcall==0.2.0
bcrypt==4.3.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
distlib==0.3.8
exceptiongroup==1.2.2
executing==2.2.0
Faker==35.2.2
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tabulate==0.9.0
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
